add prepare script

+++ module/src/main/java/com/github/kr328/clash/ProxySetup.java
@@ -17,64 +17,82 @@ class ProxySetup {
         this.baseDir = baseDir;
     }
 
-    void execOnStarted(StarterConfigure starterConfigure, ClashConfigure clashConfigure) {
-        try {
-            File script = new File(dataDir + "/mode.d/" + starterConfigure.mode + "/on-start.sh");
-            if ( !script.exists() )
-                script = new File(baseDir + "/mode.d/" + starterConfigure.mode + "/on-start.sh");
-            if ( !script.exists() ) {
-                Log.e(Constants.TAG, "Unsupported proxy mode " + starterConfigure.mode);
-                return;
-            }
-
-            HashMap<String, String> env = new HashMap<>();
-
-            if ( clashConfigure.portHttp != null )
-                env.put("CLASH_HTTP_PORT", clashConfigure.portHttp);
-            if ( clashConfigure.portSocks != null )
-                env.put("CLASH_SOCKS_PORT", clashConfigure.portSocks);
-            if ( clashConfigure.portRedir != null )
-                env.put("CLASH_REDIR_PORT", clashConfigure.portRedir);
-            if ( clashConfigure.portDns != null )
-                env.put("CLASH_DNS_PORT", clashConfigure.portDns);
-
-            env.put("CLASH_UID", Constants.CLASH_UID);
-            env.put("CLASH_GID", Constants.CLASH_GID);
-
-            exec("sh " + script.getAbsolutePath(), env);
-        } catch (IOException e) {
-            Log.e(Constants.TAG, "proxy-setup: failure", e);
+    void execOnPrepare(StarterConfigure starterConfigure, ClashConfigure clashConfigure) throws IOException {
+        File script = new File(dataDir + "/mode.d/" + starterConfigure.mode + "/on-prepare.sh");
+        if (!script.exists())
+            script = new File(baseDir + "/mode.d/" + starterConfigure.mode + "/on-prepare.sh");
+        if (!script.exists()) {
+            Log.e(Constants.TAG, "Unsupported proxy mode " + starterConfigure.mode);
+            return;
         }
+
+        HashMap<String, String> env = new HashMap<>();
+
+        if (clashConfigure.portHttp != null)
+            env.put("CLASH_HTTP_PORT", clashConfigure.portHttp);
+        if (clashConfigure.portSocks != null)
+            env.put("CLASH_SOCKS_PORT", clashConfigure.portSocks);
+        if (clashConfigure.portRedirect != null)
+            env.put("CLASH_REDIR_PORT", clashConfigure.portRedirect);
+        if (clashConfigure.portDns != null)
+            env.put("CLASH_DNS_PORT", clashConfigure.portDns);
+
+        env.put("CLASH_UID", Constants.CLASH_UID);
+        env.put("CLASH_GID", Constants.CLASH_GID);
+
+        exec("sh " + script.getAbsolutePath(), env);
     }
 
-    void execOnStop(StarterConfigure starterConfigure, ClashConfigure clashConfigure) {
-        try {
-            File script = new File(dataDir + "/mode.d/" + starterConfigure.mode + "/on-stop.sh");
-            if ( !script.exists() )
-                script = new File(baseDir + "/mode.d/" + starterConfigure.mode + "/on-stop.sh");
-            if ( !script.exists() ) {
-                Log.e(Constants.TAG, "Unsupported proxy mode " + starterConfigure.mode);
-                return;
-            }
-
-            HashMap<String, String> env = new HashMap<>();
-
-            if ( clashConfigure.portHttp != null )
-                env.put("CLASH_HTTP_PORT", clashConfigure.portHttp);
-            if ( clashConfigure.portSocks != null )
-                env.put("CLASH_SOCKS_PORT", clashConfigure.portSocks);
-            if ( clashConfigure.portRedir != null )
-                env.put("CLASH_REDIR_PORT", clashConfigure.portRedir);
-            if ( clashConfigure.portDns != null )
-                env.put("CLASH_DNS_PORT", clashConfigure.portDns);
-
-            env.put("CLASH_UID", Constants.CLASH_UID);
-            env.put("CLASH_GID", Constants.CLASH_GID);
-
-            exec("sh " + script.getAbsolutePath(), env);
-        } catch (IOException e) {
-            Log.e(Constants.TAG, "proxy-setup: failure", e);
+    void execOnStarted(StarterConfigure starterConfigure, ClashConfigure clashConfigure) throws IOException {
+        File script = new File(dataDir + "/mode.d/" + starterConfigure.mode + "/on-start.sh");
+        if (!script.exists())
+            script = new File(baseDir + "/mode.d/" + starterConfigure.mode + "/on-start.sh");
+        if (!script.exists()) {
+            Log.e(Constants.TAG, "Unsupported proxy mode " + starterConfigure.mode);
+            return;
         }
+
+        HashMap<String, String> env = new HashMap<>();
+
+        if (clashConfigure.portHttp != null)
+            env.put("CLASH_HTTP_PORT", clashConfigure.portHttp);
+        if (clashConfigure.portSocks != null)
+            env.put("CLASH_SOCKS_PORT", clashConfigure.portSocks);
+        if (clashConfigure.portRedirect != null)
+            env.put("CLASH_REDIR_PORT", clashConfigure.portRedirect);
+        if (clashConfigure.portDns != null)
+            env.put("CLASH_DNS_PORT", clashConfigure.portDns);
+
+        env.put("CLASH_UID", Constants.CLASH_UID);
+        env.put("CLASH_GID", Constants.CLASH_GID);
+
+        exec("sh " + script.getAbsolutePath(), env);
+    }
+
+    void execOnStop(StarterConfigure starterConfigure, ClashConfigure clashConfigure) throws IOException {
+        File script = new File(dataDir + "/mode.d/" + starterConfigure.mode + "/on-stop.sh");
+        if (!script.exists())
+            script = new File(baseDir + "/mode.d/" + starterConfigure.mode + "/on-stop.sh");
+        if (!script.exists()) {
+            Log.e(Constants.TAG, "Unsupported proxy mode " + starterConfigure.mode);
+            return;
+        }
+
+        HashMap<String, String> env = new HashMap<>();
+
+        if (clashConfigure.portHttp != null)
+            env.put("CLASH_HTTP_PORT", clashConfigure.portHttp);
+        if (clashConfigure.portSocks != null)
+            env.put("CLASH_SOCKS_PORT", clashConfigure.portSocks);
+        if (clashConfigure.portRedirect != null)
+            env.put("CLASH_REDIR_PORT", clashConfigure.portRedirect);
+        if (clashConfigure.portDns != null)
+            env.put("CLASH_DNS_PORT", clashConfigure.portDns);
+
+        env.put("CLASH_UID", Constants.CLASH_UID);
+        env.put("CLASH_GID", Constants.CLASH_GID);
+
+        exec("sh " + script.getAbsolutePath(), env);
     }
 
     private void exec(String command, HashMap<String, String> env) throws IOException {
@@ -85,19 +103,25 @@ class ProxySetup {
 
         Process process = builder.start();
 
-        process.getOutputStream().write(("exec " + command + " 2>&1\n").getBytes());
+        process.getOutputStream().write(("exec " + command + " 2>&1 ; exit -1\n").getBytes());
         process.getOutputStream().flush();
 
         BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
         String line;
 
-        while (( line = reader.readLine()) != null ) {
+        while ((line = reader.readLine()) != null) {
             Log.i(Constants.TAG, "proxy-setup: " + line);
         }
 
+        int result = -1;
+
         try {
-            process.waitFor();
-        } catch (InterruptedException ignored) {}
+            result = process.waitFor();
+        } catch (InterruptedException ignored) {
+        }
+
+        if (result != 0)
+            throw new IOException("Mode setup script exec failure");
 
         process.destroy();
         process.destroyForcibly();
